requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
sse-starlette==1.8.2

# OpenAI
openai==1.12.0

# LangChain for advanced RAG
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0

# HuggingFace
langchain-huggingface>=0.0.6
sentence-transformers>=2.2.0
huggingface-hub>=0.23.0

# Vector database
qdrant-client==1.7.0

# PDF processing
PyMuPDF==1.23.8
pypdf==4.0.1

# Web scraping
beautifulsoup4==4.12.3
playwright==1.41.2
requests==2.31.0

# YouTube transcription
youtube-transcript-api==1.2.4
pytube==15.0.0

# Database connectors
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pymongo==4.6.1

# Data processing
tiktoken==0.5.2
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.26.0

# Logging
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
